refactor(common): extract AvatarDisplay and getApplicantAvatar to shared
- getApplicantAvatar를 shared/hooks/avatar.ts로 이동
- AvatarDisplay 컴포넌트를 shared/components/AvatarDisplay.tsx로 분리
- profileImage → avatarEmoji → 👤 fallback 체인 공용화
